models/AutomaticTestSystem/RocmTestFramwork/requirements.txt
pytest
pytest-assume
pytest-html
pytest-timeout
pytest-repeat
allure-pytest
matplotlib
chardet
pyyaml
paddleclas
